Merged PR 32578: feat: add ecosystem crud
The aim of this PR is to create the CRUD of the ecosystem.
I have added the post and delete endpoints for the ecosystem.
In addition, I added an endpoint to get the list of skills and an endpoint to get the skills by all the ecosystems.

- feat: add ecosystem crud
- docs: update docs

Related work items: #105294